requirements.txt
playwright>=1.50.0
pytest>=7.4.0
pytest-playwright>=0.5.0
pytest-html==4.1.1
allure-pytest==2.13.2
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
schedule==1.2.1
lxml==4.9.3
